Create Marketsegmentatiom.ipynb
Effective decisions are mandatory for any company to generate good revenue. In these days competition is huge and all companies are moving forward with their own different strategies. We should use data and take a proper decision. Every person is different from one another and we don’t know what he/she buys or what their likes are. But, with the help of machine learning technique one can sort out the data and can find the target group by applying several algorithms to the dataset. Without this, It will be very difficult and no better techniques are available to find the group of people with similar character and interests in a large dataset. Here, The customer segmentation using K- Means clustering helps to group the data with same attributes which exactly helps to business the best. We are going to use elbow method to find the number of clusters and at last we visualize the data. Market Segmentation is thea process of dividing customers into groups based on common characteristics so companies can market to each group effectively and appropriately. Customers are segmented according to their similarities in behaviour and habits. Our goal is to segment their customers given the customer’s various behaviors and features, in order for the company to be able to market for each segment in a manner that is more fitting to each individual customer. Market Segmentation is one the most important applications of unsupervised learning. Using clustering techniques, companies can identify the several segments of customers allowing them to target the potential user base. In this machine learning project, we will make use of K-means clustering which is the essential algorithm for clustering unlabeled dataset.Main objective is to develop a Customer Segmentation to give recommendations like saving plans, loans, wealth management on target groups. Finally we will create a streamlit application based on the clustering technique, where we are taking the customer details and identifying which cluster the customer belongs to.